requirement.txt
tensorflow-gpu==1.15
numpy
mdtraj
mdanalysis
matplotlib
seaborn